Rename components to reduce likelihood of future confusion
